+++ noodles-vcf/src/header/info.rs
@@ -1,3 +1,0 @@
-//! VCF header info record definitions.
-
-pub mod key;