+++ noodles-vcf/src/header/record/value/map/info.rs
@@ -1,5 +1,6 @@
 //! Inner VCF header INFO map value.
 
+pub(crate) mod definition;
 mod ty;
 
 pub use self::ty::Type;
@@ -121,11 +122,11 @@ impl fmt::Display for Map<Info> {
 
 impl From<&Key> for Map<Info> {
     fn from(key: &Key) -> Self {
-        use crate::header::info::key;
-
-        let number = key::number(key).unwrap_or(Number::Count(1));
-        let ty = key::ty(key).unwrap_or(Type::String);
-        let description = key::description(key).map(|s| s.into()).unwrap_or_default();
+        let number = definition::number(key).unwrap_or(Number::Count(1));
+        let ty = definition::ty(key).unwrap_or(Type::String);
+        let description = definition::description(key)
+            .map(|s| s.into())
+            .unwrap_or_default();
 
         Self {
             inner: Info {